added authorization, otw collaboration

+++ src/api/notes/handler.js
@@ -19,11 +19,16 @@ class NotesHandler {
             const {
                 title = 'untitled', body, tags,
             } = request.payload;
+            const {
+                id: credentialId
+            } = request.auth.credentials
+
 
             const noteId = await this._service.addNote({
                 title,
                 body,
                 tags,
+                owner: credentialId
             });
 
             const response = h.response({
@@ -54,8 +59,11 @@ class NotesHandler {
         }
     }
 
-    async getNotesHandler() {
-        const notes = await this._service.getNotes()
+    async getNotesHandler(request) {
+        const {
+            id: credentialId
+        } = request.auth.credentials
+        const notes = await this._service.getNotes(credentialId)
         return {
             status: 'success',
             data: {
@@ -69,6 +77,11 @@ class NotesHandler {
             const {
                 id
             } = request.params
+            const {
+                id: credentialId
+            } = request.auth.credentials
+
+            await this._service.verifyNoteOwner(id, credentialId)
             const note = await this._service.getNoteById(id)
             return {
                 status: 'success',
@@ -104,6 +117,11 @@ class NotesHandler {
             const {
                 id
             } = request.params
+            const {
+                id: credentialId
+            } = request.auth.credentials
+
+            await this._service.verifyNoteOwner(id, credentialId)
             await this._service.editNoteById(id, request.payload)
 
             return {
@@ -137,6 +155,11 @@ class NotesHandler {
             const {
                 id
             } = request.params
+            const {
+                id: credentialId
+            } = request.auth.credentials
+
+            await this._service.verifyNoteOwner(id, credentialId)
             await this._service.deleteNoteById(id)
 
             return {
